board function separate
