+++ frontend/src/app/components/board/board.component.scss
@@ -1,4 +1,5 @@
 .gameBoard {
+    width: 300px;
     margin: .5rem;
     display: grid;
     grid-template-columns: repeat(10, auto);
